+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - restrict ssh capability to the agent socket.
 - home volume is set to the namespace name if a namespace is defined.
 - allow passing args directly to a nix run command: `podenv nix:github:podenv/devenv#emacs --version`
+- fix syscaps argument for bwrap.
 
 ## 0.1.0 (2022-05-02)